specialize npy_format_descriptor for complex type
